ai-service/services/transcription.py
import whisper

# Load the whisper model (using 'base' for a good balance of speed and accuracy)
model = whisper.load_model("base")

def transcribe_audio(file_path):
    result = model.transcribe(file_path)
    return result["text"]
