ai-service/requirements.txt
fastapi
uvicorn
groq
openai-whisper
transformers
torch
spacy
python-dotenv
pydantic
